Lua editor part1 (#66)
* Add UI Layout widget

Add HorizontalLayout widget
Add VerticalLayout widget
Add GridLayout widget
Add border to panels
Fix some shutdown related Lua issues
Improve CharacterSelect and CharacterCreation borders

* Start implementing Editor functionality to Lua scripts

Add Lua based DevTopBar editor
Add Lua based Clock Editor
Remove old ImGui based Clock Editor
Add EditorRegistry which keeps track of all Lua Editors
Improve on time drift in DayNightCycle, reducing ways for it to run faster or slower

* Fix Lua widget input sorting issue

Fix issue where Lua input events did not resolve widget ordering in the same way as the rendering

* Update Engine submodule

Update Engine submodule

+++ Source/Game-Lib/Game-Lib/ECS/Systems/UpdateDayNightCycle.h
@@ -11,8 +11,8 @@ namespace ECS::Systems
         static void Update(entt::registry& registry, f32 deltaTime);
 
         static void SetTimeToDefault(entt::registry& registry);
-        static void SetTime(entt::registry& registry, f32 time);
+        static void SetTime(entt::registry& registry, f64 time);
         static void SetSpeedModifier(entt::registry& registry, f32 speedModifier);
-        static void SetTimeAndSpeedModifier(entt::registry& registry, f32 time, f32 speedModifier);
+        static void SetTimeAndSpeedModifier(entt::registry& registry, f64 time, f32 speedModifier);
     };
 }
